gateway_grpc/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
grpcio==1.67.1
grpcio-tools==1.67.1
prometheus-client==0.21.1
pydantic==2.10.4
